+++ bonsai-ai/tests/test_brain.py
@@ -83,6 +83,11 @@ def test_brain_with_proxy():
     brain = Brain(config)
     assert brain._proxy_header() is not None
 
+    #Unset proxies
+    set_proxies(http_proxy=None,
+                https_proxy=None,
+                all_proxy=None)
+
 
 def test_brain_name_is_not_none(train_config):
     brain = Brain(train_config)
@@ -93,61 +98,6 @@ def test_brain_timeout():
     config = Config([__name__, '--network-timeout=1'])
     brain = Brain(config)
     assert brain._timeout == 1
-
-
-def test_brain_training_episode_metrics(train_config):
-    """
-      NOTE: Test does not return correct JSON according to API documentation.
-            We are verifying code paths.
-    """
-    brain = Brain(train_config)
-    metrics = brain.training_episode_metrics()
-    json.dumps(metrics)
-
-
-def test_brain_iteration_metrics(train_config):
-    """
-      NOTE: Test does not return correct JSON according to API documentation.
-            We are verifying code paths.
-    """
-    brain = Brain(train_config)
-    metrics = brain.iteration_metrics()
-    json.dumps(metrics)
-
-
-def test_brain_test_episode_metrics(train_config):
-    """
-      NOTE: Test does not return correct JSON according to API documentation.
-            We are verifying code paths.
-    """
-    brain = Brain(train_config)
-    metrics = brain.training_episode_metrics()
-    json.dumps(metrics)
-
-
-@pytest.mark.parametrize('request_errors', ['connection'], indirect=True)
-def test_brain_get_connection_error(train_config, request_errors, capsys):
-    brain = Brain(train_config)
-    brain.training_episode_metrics()
-    _, err = capsys.readouterr()
-    assert 'Unable to connect' in err
-
-
-@pytest.mark.parametrize('request_errors', ['timeout'], indirect=True)
-def test_brain_get_timeout_error(train_config, request_errors, capsys):
-    brain = Brain(train_config)
-    brain.training_episode_metrics()
-    _, err = capsys.readouterr()
-    assert 'timed out' in err
-
-
-@pytest.mark.parametrize('request_errors', ['http'], indirect=True)
-def test_brain_get_http_error(train_config, request_errors, capsys):
-    brain = Brain(train_config)
-    metrics = brain.training_episode_metrics()
-    _, err = capsys.readouterr()
-    assert 'Request failed' in err
-    assert metrics == {}
 
 
 if __name__ == '__main__':